First step at using the docs API to upload docs.
Found a few flaws with it and corrected:
- cannot return binary data in JSON so have to base64-encode it
- need to store content-type in database
- add ability to edit uploaded document

Things were getting a bit messy in api_v1 so move docs API to its own
file

Added a bunch more tests, mostly for coverage

Rolled the UI in react. It's still very problematic and ugly, and the
CSS is still not all there.

Current missing functionality:
- enforcing max file size limits
- doing previews correctly
- enforcing content-type both on server and client
- enforcing number of files per user
- keeping track of amount of space used by the files
- better transition between 'new' and 'existing' view